move pg lib into model module and import sqls submodule
