+++ crates/pg/src/sqls.rs
@@ -1,0 +1,9 @@
+pub mod account;
+pub mod approval;
+pub mod balance;
+pub mod common;
+pub mod cte;
+pub mod profile;
+pub mod rule_instance;
+pub mod transaction;
+pub mod transaction_item;
