+++ services/rule/Cargo.toml
@@ -13,7 +13,6 @@ tower-http = "0.4.4"
 tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 types = { path = "../../crates/types" }
-sqls = { path = "../../crates/sqls" }
 pg = { path = "../../crates/pg" }
 chrono = "0.4.23"
 nanoid = "0.4.0"
